+++ Utopia/libs/librabbitmq/amqp_openssl.c
@@ -86,9 +86,9 @@ amqp_ssl_socket_send(void *base,
 
   /* This will only return on error, or once the whole buffer has been
    * written to the SSL stream. See SSL_MODE_ENABLE_PARTIAL_WRITE */
-  res = SSL_write(self->ssl, buf, len);
+  res = SSL_write(self->ssl, buf, (int)len);
   if (0 >= res) {
-    self->internal_error = SSL_get_error(self->ssl, res);
+    self->internal_error = SSL_get_error(self->ssl, (int)res);
     /* TODO: Close connection if it isn't already? */
     /* TODO: Possibly be more intelligent in reporting WHAT went wrong */
     switch (self->internal_error) {
@@ -151,9 +151,9 @@ amqp_ssl_socket_recv(void *base,
   ERR_clear_error();
   self->internal_error = 0;
 
-  received = SSL_read(self->ssl, buf, len);
+  received = SSL_read(self->ssl, buf, (int)len);
   if (0 >= received) {
-    self->internal_error = SSL_get_error(self->ssl, received);
+    self->internal_error = SSL_get_error(self->ssl, (int)received);
     switch(self->internal_error) {
     case SSL_ERROR_ZERO_RETURN:
       received = AMQP_STATUS_CONNECTION_CLOSED;
@@ -236,7 +236,7 @@ amqp_ssl_socket_open(void *base, const char *host, int port, struct timeval *tim
 
   self->ssl = SSL_new(self->ctx);
   if (!self->ssl) {
-    self->internal_error = ERR_peek_error();
+    self->internal_error = (int)ERR_peek_error();
     status = AMQP_STATUS_SSL_ERROR;
     goto exit;
   }
@@ -266,7 +266,7 @@ amqp_ssl_socket_open(void *base, const char *host, int port, struct timeval *tim
 
   result = SSL_get_verify_result(self->ssl);
   if (X509_V_OK != result) {
-    self->internal_error = result;
+    self->internal_error = (int)result;
     status = AMQP_STATUS_SSL_PEER_VERIFY_FAILED;
     goto error_out3;
   }
@@ -449,7 +449,7 @@ amqp_ssl_socket_set_key_buffer(amqp_socket_t *base,
   if (1 != status) {
     return AMQP_STATUS_SSL_ERROR;
   }
-  buf = BIO_new_mem_buf((void *)key, n);
+  buf = BIO_new_mem_buf((void *)key, (int)n);
   if (!buf) {
     goto error;
   }
